agent-core/requirements.txt
fastapi
uvicorn
pyautogui
google-generativeai
pillow
python-dotenv
